Python testing: Move code out of main file
This file is becoming large, so I moved out some of the code that
is logically similar. I also abstraceted the setup into a separate
function. The reason for this: Right now, all these tests are run
in the CI against all clusters. All cluster is a special case and
it is not clear that it is ever going to be spec conformant.
I would like to split that test out, enable the error and start
running it against individual example apps as they get corrected.

There are no code changes in this PR, just moves and adjustments
of the imports.
