+++ src/python_testing/TestMatterTestingSupport.py
@@ -25,9 +25,9 @@ from chip.tlv import uint
 from matter_testing_support import (MatterBaseTest, async_test_body, compare_time, default_matter_test_main,
                                     get_wait_seconds_from_set_time, parse_pics, type_matches, utc_time_in_matter_epoch)
 from mobly import asserts, signals
-from TC_DeviceBasicComposition import (TagProblem, create_device_type_list_for_root, create_device_type_lists,
-                                       find_tag_list_problems, find_tree_roots, get_all_children, get_direct_children_of_root,
-                                       parts_list_cycles, separate_endpoint_types)
+from taglist_and_topology_test_support import (TagProblem, create_device_type_list_for_root, create_device_type_lists,
+                                               find_tag_list_problems, find_tree_roots, get_all_children,
+                                               get_direct_children_of_root, parts_list_cycles, separate_endpoint_types)
 
 
 def get_raw_type_list():
